requirements.txt
faster-whisper==1.0.3
resemblyzer
librosa==0.10.1
soundfile
scikit-learn
scipy
numpy
gdown
pandas
plotly
reportlab
python-docx
PyYAML
torch
google-generativeai
rich
flask
flask-cors

